+++ src/com/genir/aitweaks/utils/ShipStash.kt
@@ -1,5 +1,6 @@
 package com.genir.aitweaks.utils
 
+import com.fs.starfarer.api.Global
 import com.fs.starfarer.api.combat.ShipAPI
 import com.genir.aitweaks.features.shipai.ai.Maneuver
 
@@ -8,22 +9,26 @@ private const val stashKey = "aitweaks_ship_stash"
 const val maxAge = 0.5f
 
 class ShipStash(ship: ShipAPI) {
-    /** Enemy target the ship is attacking. */
+    /** Enemy target the ship is attacking, selected by Custom AI. */
     var attackTarget: ShipAPI? = null
-        get() = if (frameTracker - attackTargetT < maxAge) field else null
+        get() = if (elapsedTime() - attackTargetT < maxAge) field else null
         set(value) {
-            field = value; attackTargetT = frameTracker
+            field = value; attackTargetT = elapsedTime()
         }
 
-    /** Current instance of Maneuver AI controlling the ship.*/
+    /** Current instance of Custom AI Maneuver controlling the ship.*/
     var maneuverAI: Maneuver? = null
-        get() = if (frameTracker - maneuverAIT < maxAge) field else null
+        get() = if (elapsedTime() - maneuverAIT < maxAge) field else null
         set(value) {
-            field = value; maneuverAIT = frameTracker
+            field = value; maneuverAIT = elapsedTime()
         }
 
     private var attackTargetT = 0f
     private var maneuverAIT = 0f
+
+    private fun elapsedTime(): Float {
+        return Global.getCombatEngine().getTotalElapsedTime(false)
+    }
 }
 
 val ShipAPI.aitStash: ShipStash